Group school year commands, queries, and entities into a dedicated namespace.
