+++ Application/EdFi.Ods.AdminApp.Management.Tests/AssertionExtensions.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using EdFi.Ods.AdminApp.Management.Database.Ods;
+using EdFi.Ods.AdminApp.Management.Database.Ods.SchoolYears;
 using FluentValidation;
 using FluentValidation.Results;
 using Shouldly;
