+++ Application/EdFi.Ods.AdminApp.Management/Database/Ods/SchoolYears/SchoolYearType.cs
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+namespace EdFi.Ods.AdminApp.Management.Database.Ods.SchoolYears
+{
+    public class SchoolYearType
+    {
+        public short SchoolYear { get; set; }
+        public string SchoolYearDescription { get; set; }
+        public bool CurrentSchoolYear { get; set; }
+    }
+}
